+++ app/src/main/java/com/bguneys/app652020/note/NoteListFragment.kt
@@ -1,5 +1,6 @@
 package com.bguneys.app652020.note
 
+import androidx.appcompat.app.AlertDialog
 import android.content.Context
 import android.os.Bundle
 import android.text.TextUtils
@@ -103,11 +104,27 @@ class NoteListFragment : Fragment() {
 
             // Called when a ViewHolder is swiped by the user. Used for swipe to delete functionality
             override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
-                val position = viewHolder.adapterPosition
-                val selectedFolder = adapter.noteList.get(position)
-                noteViewModel.deleteByFolderTitleAndNoteTitle(
-                    selectedFolder.folderTitle,
-                    selectedFolder.noteTitle!!)
+
+                //show dialog to ensure deleting note
+                val dialogBuilder = AlertDialog.Builder(context!!)
+                dialogBuilder.setMessage("Delete folder?")
+                    .setCancelable(false)
+                    .setPositiveButton("Delete", {dialog, id ->
+                        //delete note
+                        val position = viewHolder.adapterPosition
+                        val selectedFolder = adapter.noteList.get(position)
+                        noteViewModel.deleteByFolderTitleAndNoteTitle(
+                            selectedFolder.folderTitle,
+                            selectedFolder.noteTitle!!)
+
+                        })
+                    .setNegativeButton("Cancel", { dialog, id ->
+                        dialog.dismiss() //do nothing and dismiss the dialog
+                        adapter.notifyDataSetChanged() //revert swipe action when cancelled
+                    })
+
+                val alert = dialogBuilder.create()
+                alert.show()
             }
         }
